# openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/auth/profile:
    get:
      description: Retrieves the current user profile from Clerk
      operationId: getProfile
      parameters: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get user profile
      tags:
        - Auth
  /auth/auth/session-claims:
    get:
      description: Retrieves the current session JWT claims
      operationId: getSessionClaims
      parameters: []
      responses:
        '200':
          description: Session claims retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionClaimsResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get session claims
      tags:
        - Auth
...
  /areas:
    post:
      operationId: AreasController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAreaDto'
      responses:
        '201':
          description: Area created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponseDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
      security:
        - BearerAuth: []
      summary: Create a new area
      tags:
        - Areas
    get:
      operationId: AreasController_findAll
      parameters:
        - name: location_id
          required: false
          in: query
          description: Filter by location ID
          schema:
            format: uuid
            type: string
        - name: parent_id
          required: false
          in: query
          description: Filter by parent area ID
          schema:
            format: uuid
            type: string
        - name: root_only
          required: false
          in: query
          description: Filter root areas only (no parent)
          schema:
            type: boolean
        - name: is_active
          required: false
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: include_children
          required: false
          in: query
          description: Include children in response (hierarchical)
          schema:
            type: boolean
      responses:
        '200':
          description: List of areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AreaResponseDto'
        '401':
          description: Unauthorized
      security:
        - BearerAuth: []
      summary: List areas with optional filtering
      tags:
        - Areas
  /areas/{id}:
    get:
      operationId: AreasController_findById
      parameters:
        - name: id
          required: true
          in: path
          description: Area ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Area found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponseDto'
        '401':
          description: Unauthorized
        '404':
          description: Area not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - BearerAuth: []
      summary: Get area by ID
      tags:
        - Areas
    put:
      operationId: AreasController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Area ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAreaDto'
      responses:
        '200':
          description: Area updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponseDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
        '404':
          description: Area not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - BearerAuth: []
      summary: Update an area
      tags:
        - Areas
    delete:
      operationId: AreasController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Area ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Area deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Area not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - BearerAuth: []
      summary: Delete an area
      tags:
        - Areas
  /areas/{id}/children:
    get:
      operationId: AreasController_findByIdWithChildren
      parameters:
        - name: id
          required: true
          in: path
          description: Area ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Area with children
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponseDto'
        '401':
          description: Unauthorized
        '404':
          description: Area not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - BearerAuth: []
      summary: Get area with children
      tags:
        - Areas
...
info:
  title: LibreStock Inventory API
  description: REST API for LibreStock Inventory Management System
  version: 1.0.0
  contact:
    name: API Support
    url: ''
    email: ''
  license:
    name: MIT
    url: ''
...
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.librestock-inventory.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token
  schemas:
    ProfileResponseDto:
      type: object
      properties: {}
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
      required:
        - error
    SessionClaimsResponseDto:
      type: object
      properties:
        user_id:
          type: string
        session_id:
          type: string
        expires_at:
          type: number
          format: int64
        issued_at:
          type: number
          format: int64
      required:
        - user_id
        - session_id
        - expires_at
        - issued_at
...
    CreateAreaDto:
      type: object
      properties:
        location_id:
          type: string
          description: Location ID
          format: uuid
        parent_id:
          type: string
          description: Parent area ID (for nested areas)
          format: uuid
        name:
          type: string
          description: Area name
          example: Zone A
        code:
          type: string
          description: Area code (short identifier)
          example: A1
        description:
          type: string
          description: Area description
        is_active:
          type: boolean
          description: Whether the area is active
          default: true
      required:
        - location_id
        - name
    AreaResponseDto:
      type: object
      properties:
        created_at:
          format: date-time
          type: string
          description: Creation timestamp
        updated_at:
          format: date-time
          type: string
          description: Last update timestamp
        id:
          type: string
          description: Unique identifier
          format: uuid
        location_id:
          type: string
          description: Location ID
          format: uuid
        parent_id:
          type: object
          description: Parent area ID
          format: uuid
          nullable: true
        name:
          type: string
          description: Area name
        code:
          type: string
          description: Area code
        description:
          type: string
          description: Area description
        is_active:
          type: boolean
          description: Whether the area is active
        children:
          description: Child areas
          type: array
          items:
            $ref: '#/components/schemas/AreaResponseDto'
      required:
        - created_at
        - updated_at
        - id
        - location_id
        - parent_id
        - name
        - code
        - description
        - is_active
    UpdateAreaDto:
      type: object
      properties:
        parent_id:
          type: object
          description: Parent area ID (for nested areas)
          format: uuid
        name:
          type: string
          description: Area name
          example: Zone A
        code:
          type: string
          description: Area code (short identifier)
          example: A1
        description:
          type: string
          description: Area description
        is_active:
          type: boolean
          description: Whether the area is active
...